if_infiniband: Fix issue preventing BPF writes to IPoIB interfaces
The if_infiniband driver is currently unable able to support broadcast
frame injection via BPF. This is because the infiniband_output function
fragments the injected frame header due to a mismatch in header size
between the actual infiniband frame and the BPF registered device type
(EN10MB).

The EN10MB type registration causes BPF to interpret the first 14 bytes
of any frames written to the BPF device via the write system call as an
Ethernet header. BPF extracts this header from the frame mbuf and passes
it to the if_infiniband output function in the ro->ro_prepend member.

The if_infiniband driver attempts to inject this prepend data back into
the packet, however it does so after first pre-allocating the infiniband
header size to the front of the mbuf. This causes a 10 byte gap between
the data supplied in ro_prepend from BPF and the remaining packet data.

This change addresses the problem by pre-allocating the variable header
length tracked by hlen, rather always pre-allocating a static infiniband
header structure. While BPF has fragmented the IB header, this will
reassemble it at the correct location prior to sending it via the
if_transmit call without requiring native IB support to be added to BPF.
It also corrects any invalid mbuf flags injected by BPF due to
misinteretation of the IB header as an Ethernet header.

The rationale for requiring a functional BPF write path is that it is
necessary to support DHCP over IPoIB. The DHCP protocol needs the
ability to inject link-layer Broadcasts to precise interfaces, which
is satisfied by BPF.

As BPF is still interpreting headers as Ethernet, users of BPF for IPoIB
devices will need to set the BIOCSHDRCMPLT ioctl to 1 to avoid header
mangling until native headers support is incorporated in BPF.

Signed-off-by: Nicholas Dance <ndance@vdura.com>